requirements_test.txt
givenergy-modbus==0.10.1
pytest-homeassistant-custom-component==0.8.1
